add breakpoint

+++ hfq2qfq.py
@@ -15,8 +15,12 @@ def _doqfq(code):
     logging.info("qfq for %s."%code)
     hfqDF = pd.read_sql('select * from t_daily_hfq_stock where code="%s"'%code, utils.getEngine()).set_index('date').sort_index(ascending=False)
 
-    factor = pd.read_sql('select * from t_daily_fqFactor where code="%s" and date="%s"'%(code, hfqDF.head(1).index.values[0]), utils.getEngine())['factor'].values[0]
+    factors = pd.read_sql('select * from t_daily_fqFactor where code="%s" and date="%s"'%(code, hfqDF.head(1).index.values[0]), utils.getEngine())
+    # off market
+    if factors.empty:
+        return
 
+    factor = factors['factor'].values[0]
 
     rt = ts.get_realtime_quotes(code)
     if rt is None:
@@ -43,15 +47,17 @@ def _doqfq(code):
         print factor
         print _rate
 
-def _hfq2qfq():
+def _hfq2qfq(beginCode=None):
 
     codes = pd.read_sql('select code from t_stock_basics', utils.getEngine())
     if conf.DEBUG:
         codes = codes[:1]
 
-    if not conf.DEBUG:
+    if not conf.DEBUG and beginCode is None:
         utils.executeSQL('delete from t_daily_qfq_stock')
 
+    if beginCode is not None:
+        codes = codes[codes['code']>=beginCode]
     codes['code'].map(_doqfq)
     
 
@@ -59,9 +65,10 @@ if '__main__' == __name__:
 
     parser = argparse.ArgumentParser()
     parser.add_argument("-p", "--production", action="store_true", help='''defalt is in debug mode, which only plays a little''')
+    parser.add_argument("-b", "--beginCode", type=str, help='''starts with this code''')
     args = parser.parse_args()
 
     if args.production:
         conf.DEBUG = False
 
-    _hfq2qfq()
+    _hfq2qfq(args.beginCode)
